add performance optimization for ggml_compute_forward_flash_ff_gated_f32
dot(a[:,i1], w1[:,iw20]) and dot(a[:,i1], w3[:,iw20]) where recomputed ne0 times!
swap loops i0 and iw20, so that i0 is inner loop.
but now we can't add complete dot sum for dst[i0,i1] in inner loop.
instead we need to accumulate directly into dst[i0,i1].
for this we must guarantee that no two threads share the same i1.

printf output of i1 bounds for each thread showed that this is the case.